Refactor hashable and findable into calculable
