+++ lib/dependents.rb
@@ -1,6 +1,5 @@
 require_relative 'game.rb'
 require_relative 'team.rb'
 require_relative 'game_teams.rb'
-require_relative 'hashable.rb'
 require_relative 'calculable.rb'
 require_relative 'season.rb'
